pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "roweeder"
version = "1.0.0"
description = "RoWeeder: A deep learning model for weed detection and crop row segmentation"
readme = "README.md"
license = {text = "LICENSE"}
authors = [
    {name = "Pasquale De Marinis", email = "pasquale.demarinis@unicatt.it"}
]
requires-python = ">=3.8"

dependencies = [
    "torch>=1.12.0",
    "torchvision>=0.13.0",
    "transformers>=4.25.0",
    "accelerate>=0.20.0",
    "wandb>=0.13.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "scikit-image>=0.19.0",
    "opencv-python>=4.5.0",
    "Pillow>=8.3.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "tqdm>=4.62.0",
    "torchmetrics>=0.10.0",
    "ptflops>=0.7.0",
    "safetensors>=0.3.0",
    "albumentations>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2",
    "pytest-cov>=2.12",
    "black>=21.5b0",
    "flake8>=3.9",
    "isort>=5.9",
]

[tool.setuptools]
packages = ["roweeder"]

[tool.setuptools.package-data]
roweeder = ["**/*.py"]
